Test de stress. Se agregan opciones

+++ stress/test.sh
@@ -1,20 +1,76 @@
 #!/bin/bash
+
+# TODO: Agregar getopts para distintos parametros (cantidad de movidas, tipos de juego, procesos)
+#
+
+function   usage(){
+    cat  <<EOF
+
+usage: $0  options
+
+Options:
+  -c      Iniciar la base (no iniciar por defecto)
+  -g n    Cantidad de juegos en paralelo (default $GAMES)
+  -h s    Hostname (default: localhost)
+  -k n    Workers (experimental)
+  -m n    Cantidad de movidas maxima por partido (default $MOVIDAS)
+  -n      Usar valgrind (no usar por defecto)
+  -p n    Puerto (default: $PORT)
+  -s      Iniciar el servidor qgserverd (no iniciar por defecto)
+  -v      Verbose
+  -w      Watch!
+
+EOF
+}
+
 PATHQS=./src
 DBFILE=./stress/qgserver.db
 VERBOSE=
+WORKERS=
+HOST=localhost
 PORT=8015
+GAMES=30
+MOVIDAS=200
+
+if [ x$1 == x ]; then
+    usage; exit 0;
+fi
+
+
+
+
+
+
+while getopts "vg:k:wh:p:scm:n" OPTION; do
+    case $OPTION in
+      c) INITDB=1;;
+      g) GAMES=$OPTARG;;
+      h) HOST=$OPTARG;;
+      k) WORKERS=-w $OPTARG;;
+      m) MOVIDAS=$OPTARG;;
+      n) VALGRIND=1;;
+      p) PORT=$OPTARG;;
+      s) STARTQG=1;;
+      v) VERBOSE=-v;;
+      w) WATCH=1;;
+      *) usage; exit 0;
+    esac
+done
+
 
 
 . ./stress/functions.sh
 
 
 # Primeramente, cargo una base de datos vacia
-rm $DBFILE
-echo "Creando archivo"
-$PATHQS/qgserver-tool -c $DBFILE
-ret=$?
-if [ $ret != 0 ]; then
-    exit 1;
+if [ x$INITDB == x1 ]; then
+    rm $DBFILE
+    echo "Creando archivo"
+    $PATHQS/qgserver-tool -c $DBFILE
+    ret=$?
+    if [ $ret != 0 ]; then
+        exit 1;
+    fi
 fi
 
 # el directorio de archivos temporales.
@@ -23,11 +79,20 @@ if [ ! -e ./stress/tmp ]; then
 fi
 
 # Voy a intentar ejecutar el servicio web
-echo "Iniciando servidor"
-$PATHQS/qgserverd -d $DBFILE -p $PORT $VERBOSE > ./stress/tmp/salida.txt &
-PID=$!
+if [ x$STARTQG == x1 ]; then
+    if [ x$VALGRIND == x1 ]; then
+        echo "Iniciando servidor valgrind"
+        valgrind  --leak-check=full  $PATHQS/qgserverd -d $DBFILE -p $PORT $VERBOSE $WORKERS &> ./stress/tmp/salida.txt &
+        PID=$!
+        sleep 5 # le doy un segundo como para que levante
+    else
+        echo "Iniciando servidor"
+        $PATHQS/qgserverd -d $DBFILE -p $PORT $VERBOSE $WORKERS > ./stress/tmp/salida.txt &
+        PID=$!
+        sleep 1 # le doy un segundo como para que levante
+    fi
+fi
 
-sleep 1 # le doy un segundo como para que levante
 
 sess=$(sesslogin)
 
@@ -35,9 +100,8 @@ function   juega_hasta_200(){
     tipojuego=`selecttipojuego`
     num=$1
     echo $tipojuego
-    movidas=200
     game=`creajuego $tipojuego`
-    for i in `seq 1 $movidas`; do
+    for i in `seq 1 $MOVIDAS`; do
         move=`selectposibles $game`
         mueve $move
 
@@ -64,7 +128,7 @@ function   juega_hasta_200(){
 function  watch_registraciones(){
     sleep 1
     while [ true ]; do 
-        # clear
+        clear
         registraciones 
         sleep 2
     done
@@ -73,19 +137,15 @@ function  watch_registraciones(){
 
 
 waiting=""
-for i in `seq 1 30`; do
+for i in `seq 1 $GAMES`; do
     juega_hasta_200 $i > ./stress/tmp/$i &
     waiting="$waiting $!"
 done
 
-echo "Inicio registraciones"
-watch_registraciones &
-PIDR=$!
-wait  $waiting
+if [ x$WATCH == x1 ]; then
+    watch_registraciones &
+fi
 
-echo "Fin de todos los procesos, cancelo $PID $PIDR"
-kill -9 $PIDR
-wait $PIDR
-kill -2 $PID
+wait  $waiting
 wait
 exit 0
